+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - Fixed the subunit_to_unit values of the CLP and KRW currencies
  - Add option for `disambiguate` symbols for formatting
  - Fixed the subunit_to_unit values of the VND currency
+ - Fix formatting of NGN - show symbol before amount
 
 ## 6.2.1
  - Ensure set is loaded